move the helper function to util.py
